Traspaso de chats, mostrar chats y mensajes correctamente. Seguridad de chats. Opciones de chat de admin + seguridad en las operaciones de admin (no había).

+++ modelo/php/admin/chatActions.php
@@ -2,20 +2,39 @@
 
 include "../../../common/php/connect.php";
 
-if(isset($_GET["controlChat"])) { 
-    $chatcontrol = $_GET['controlChat'];
-    $cid = $_GET['cid'];
-    if($chatcontrol != null){
-        if($chatcontrol == 1){
-            header("Location: ../../../vista/php/chat.php?cid=$cid");
-        }else if($chatcontrol == 2){
-            $sql = "DELETE FROM chats
-            WHERE id = '".$cid."';";
+if (isset($_SESSION['id'])) {
+    $sqlc = "SELECT
+    id_tipo
+    FROM usuarios
+    WHERE id = " . $_SESSION['id'];
 
-            $mysqliresult = $conn->query($sql);
-            header("Location: ../../../vista/php/admin.php#manage-chats");
-        }
+    $resultc = $conn->query($sqlc);
+    $rowc = $resultc->fetch_assoc();
+
+    if ($rowc['id_tipo'] == 1) {
+        if(isset($_GET["controlChat"])) { 
+            $chatcontrol = $_GET['controlChat'];
+            $cid = $_GET['cid'];
+            if($chatcontrol != null){
+                if($chatcontrol == 1){
+                    header("Location: ../../../vista/php/chat.php?cid=$cid");
+                }else if($chatcontrol == 2){
+                    $sql = "DELETE FROM chats
+                    WHERE id = '".$cid."';";
+        
+                    $mysqliresult = $conn->query($sql);
+                    header("Location: ../../../vista/php/admin.php#manage-chats");
+                }
+            }
+        }   
+    } else {
+        header("Location: ../../../common/php/disconnect.php");
     }
+
+} else {
+    header("Location: ../../../vista/php/login.php");
 }
+
+$conn->close();
 
 ?>
